Hero List
Created a list of heroes with mock data. Used *ngFor to display the
list. Used a private CSS stylings for the Component for styling the
list. Used *ngIf to hide components that weren't valid yet. Used (click)
to select a list item. Created the onSelect function to work with
(click). Used the [class.selected] styling to style the item on the list
that is selected.