Routing
Added Routes. Created new DashboardComponent which offers a different
way to link to details of a hero.
Decoupled the Details from the HeroesComponent so it can stand on its own
and be used from multiple places.
Added RoutingModule to setup the routes.
Made DashboardComponent the default route.
Added a 'go back' button to the details page to go back to whichever
location 'called' it.
Added a new service method to get a specific hero based on id. Still
using mock data at this time.